:construction: Wip on the move of the DTO in server
