+++ client/CMakeLists.txt
@@ -1,16 +1,13 @@
 add_executable(r-type_client)
 
 target_link_libraries(r-type_client PUBLIC StellarForge StellarForgeGraphics StellarForgePhysics StellarForgeCommon sfml::sfml glm::glm PRIVATE rtype::network)
-add_subdirectory(../network/protocol)
 
 target_sources(r-type_client
         PUBLIC
         ${CMAKE_CURRENT_SOURCE_DIR}/../assets/objects/scripts/Background.hpp
         ${CMAKE_CURRENT_SOURCE_DIR}/../assets/objects/scripts/NetworkManager.hpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/utils/BinaryVector.hpp
         PRIVATE
         ${CMAKE_CURRENT_SOURCE_DIR}/main.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/../assets/objects/scripts/Background.cpp
         ${CMAKE_CURRENT_SOURCE_DIR}/../assets/objects/scripts/NetworkManager.cpp
-        ${CMAKE_CURRENT_SOURCE_DIR}/utils/BinaryVector.cpp
 )
